一覧画面にデータの数に応じてCellを反映させる(Vに集中しているためVとCにリファクタリングする必要あり)
